fix: login/logout redirects, runtime URL resolution, SDK host, live profile name
Bugs found during browser testing of the deployment branch across dev,
docker-dev, and docker-prod:

1. 405 on /login after logout — the middleware redirected unauthenticated
   requests with NextResponse.redirect's default 307, which preserves the
   method, so a POST to a gated route (the logout form once the session lapsed,
   any plugin form) became POST /login → 405. Now 303 (See Other) so the browser
   GETs /login. The logout route also redirects to SOVEREIGN_AUTH_PUBLIC_URL
   (browser-reachable) instead of the internal SOVEREIGN_AUTH_URL.

2. Post-login redirect went to :3000 in docker-prod — the auth login/register
   pages were 'use client' and read NEXT_PUBLIC_RUNTIME_URL directly, which Next
   inlines at build time; the image builds without .env, freezing it to the
   localhost:3000 default. Now resolved server-side at request time
   (apps/auth/src/runtime-url.ts, computed-key read to dodge inlining) and passed
   as a prop to new client LoginForm/RegisterForm. Prod compose defaults
   NEXT_PUBLIC_RUNTIME_URL/AUTH_BASE_URL to the host ports, and .env.example
   comments out the URL vars so dev values don't leak into prod.

3. Logout flickered without navigating — the CSP form-action was 'self' only,
   so the browser blocked the logout form's cross-origin 303 to the auth login
   page. form-action now includes the browser-facing auth origin
   (authPublicOrigin() → buildContentSecurityPolicy).

4. "no runtime host is registered" when changing the display name — the SDK host
   was a module-level singleton; Next bundles instrumentation, route handlers,
   and server actions separately (and HMR re-evaluates modules), so provideHost()
   set it on one instance while the account server action read null from another.
   The host is now stored on globalThis under a Symbol.for key, shared across all
   module instances in the process. (@sovereignfs/sdk 1.3.1)

5. Updated name didn't appear in the account dialog until reopened — the profile
   page read the name from the session-cache headers, stale within the same
   action cycle. It now reads the authoritative value from
   get-session?disableCookieCache=true (plain get-session honours the same cache).
   (@sovereignfs/plugin-account 0.4.1)

Versions: @sovereignfs/sdk 1.3.1, @sovereignfs/plugin-account 0.4.1,
apps/auth 0.5.3, runtime 0.15.1. Platform stays frozen at 0.6.0.
