+++ .env.example
@@ -22,10 +22,19 @@ AUTH_INVITE_ONLY=false
 # AUTH_TRUSTED_ORIGINS=http://auth:3001
 
 # --- Runtime (runtime/) ---
-# Where the runtime reaches the auth server for /api/verify, and where the auth
-# server sends users back after login.
-SOVEREIGN_AUTH_URL=http://localhost:3001
-NEXT_PUBLIC_RUNTIME_URL=http://localhost:3000
+# Where the runtime reaches the auth server (server-to-server, e.g. /api/verify)
+# and where the auth server sends users back after login.
+#
+# Leave both UNSET for local use: native `pnpm dev` falls back to the localhost
+# defaults below, and the Compose files inject the right values per environment
+# (dev → :3000/:3001, prod → :4000/:4001 on the host). Setting them here would
+# override those per-environment defaults for ALL environments at once.
+#
+# For a REAL production deployment, set NEXT_PUBLIC_RUNTIME_URL to your public
+# runtime URL (e.g. https://example.com). SOVEREIGN_AUTH_URL stays the internal
+# address (the Compose files set it to http://auth:3001).
+# SOVEREIGN_AUTH_URL=http://localhost:3001
+# NEXT_PUBLIC_RUNTIME_URL=http://localhost:3000
 # Shared secret for runtime→auth admin API calls (Console user management).
 # No default — the auth server and runtime both refuse to start without it.
 # Generate one with:  openssl rand -base64 32
