+++ apps/auth/app/register/page.tsx
@@ -1,80 +1,9 @@
-'use client';
+import { runtimePublicUrl } from '@/src/runtime-url';
+import { RegisterForm } from './register-form';
 
-import { type FormEvent, useState } from 'react';
-import Link from 'next/link';
-import { Button, Input } from '@sovereignfs/ui';
-import { authClient } from '@/src/auth-client';
-import styles from '../auth.module.css';
-
-const RUNTIME_URL = process.env.NEXT_PUBLIC_RUNTIME_URL ?? 'http://localhost:3000';
-
+// Server component: resolves the runtime URL at request time and hands it to the
+// client form (so the post-registration redirect uses the deployment's real
+// runtime origin, not a build-time-frozen value).
 export default function RegisterPage() {
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [error, setError] = useState<string | null>(null);
-  const [loading, setLoading] = useState(false);
-
-  async function onSubmit(event: FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    setLoading(true);
-    setError(null);
-    const result = await authClient.signUp.email({ name, email, password });
-    setLoading(false);
-    if (result.error) {
-      setError(result.error.message ?? 'Registration failed.');
-      return;
-    }
-    window.location.href = RUNTIME_URL;
-  }
-
-  return (
-    <main className={styles.page}>
-      <div className={styles.card}>
-        <h1 className={styles.title}>Create your account</h1>
-        <form className={styles.form} onSubmit={onSubmit}>
-          <label className={styles.field}>
-            <span className={styles.label}>Name</span>
-            <Input
-              autoComplete="name"
-              required
-              value={name}
-              onChange={(e) => setName(e.target.value)}
-            />
-          </label>
-          <label className={styles.field}>
-            <span className={styles.label}>Email</span>
-            <Input
-              type="email"
-              autoComplete="email"
-              required
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-            />
-          </label>
-          <label className={styles.field}>
-            <span className={styles.label}>Password</span>
-            <Input
-              type="password"
-              autoComplete="new-password"
-              required
-              minLength={8}
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-            />
-          </label>
-          {error ? <p className={styles.error}>{error}</p> : null}
-          <Button type="submit" disabled={loading}>
-            {loading ? 'Creating account…' : 'Create account'}
-          </Button>
-        </form>
-        <p className={styles.footer}>
-          Already have an account?{' '}
-          <Link className={styles.link} href="/login">
-            Sign in
-          </Link>
-        </p>
-      </div>
-    </main>
-  );
+  return <RegisterForm runtimeUrl={runtimePublicUrl()} />;
 }
