+++ apps/auth/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sovereignfs/auth",
-  "version": "0.5.2",
+  "version": "0.5.3",
   "private": true,
   "type": "module",
   "scripts": {
