+++ plugins/account/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sovereignfs/plugin-account",
-  "version": "0.4.0",
+  "version": "0.4.1",
   "private": true,
   "type": "module",
   "dependencies": {
