+++ runtime/app/api/account/logout/route.ts
@@ -1,7 +1,15 @@
 import { NextResponse } from 'next/server';
 import { sdk } from '@sovereignfs/sdk';
 
-const AUTH_URL = process.env.SOVEREIGN_AUTH_URL ?? 'http://localhost:3001';
+// Browser-facing auth URL for the post-logout redirect — must be reachable from
+// the browser. In Docker, SOVEREIGN_AUTH_URL is the internal service name
+// (http://auth:3001) which the browser cannot resolve, so prefer the public URL
+// (same precedence as runtime/app/login/route.ts). sdk.auth.signOut() makes its
+// own server-to-server call using SOVEREIGN_AUTH_URL internally.
+const AUTH_PUBLIC_URL =
+  process.env.SOVEREIGN_AUTH_PUBLIC_URL ??
+  process.env.SOVEREIGN_AUTH_URL ??
+  'http://localhost:3001';
 
 /**
  * Self sign-out (AUTH-02 / ACC-11). A POST target so both the shell avatar menu
@@ -19,7 +27,7 @@ const AUTH_URL = process.env.SOVEREIGN_AUTH_URL ?? 'http://localhost:3001';
 export async function POST(): Promise<Response> {
   await sdk.auth.signOut();
 
-  const res = NextResponse.redirect(`${AUTH_URL}/login?signedout=1`, 303);
+  const res = NextResponse.redirect(`${AUTH_PUBLIC_URL}/login?signedout=1`, 303);
   // Drop both cache-cookie variants (the `__Secure-` one only unsets with Secure).
   res.cookies.set('better-auth.session_data', '', { maxAge: 0, path: '/' });
   res.cookies.set('__Secure-better-auth.session_data', '', {
